integration: extract test_token_calculation()
The test is extracted into `statements/prepared.rs`.
