+++ scylla/tests/integration/statements/prepared.rs
@@ -1,8 +1,10 @@
-use scylla::{
-    response::{PagingState, PagingStateResponse},
-    routing::{partitioner::PartitionerName, Token},
-    statement::Statement,
-};
+use scylla::client::session::Session;
+use scylla::response::{PagingState, PagingStateResponse};
+use scylla::routing::partitioner::PartitionerName;
+use scylla::routing::Token;
+use scylla::serialize::row::SerializeRow;
+use scylla::statement::prepared::PreparedStatement;
+use scylla::statement::Statement;
 
 use crate::utils::{
     create_new_session_builder, scylla_supports_tablets, setup_tracing, unique_keyspace_name,
@@ -253,4 +255,162 @@ async fn test_prepared_partitioner() {
         prepared_statement_for_cdc_log.get_partitioner_name(),
         &PartitionerName::CDC
     );
+}
+
+#[tokio::test]
+async fn test_token_calculation() {
+    setup_tracing();
+    let session = create_new_session_builder().build().await.unwrap();
+    let ks = unique_keyspace_name();
+
+    session.ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks)).await.unwrap();
+    session.use_keyspace(ks.as_str(), true).await.unwrap();
+
+    #[allow(clippy::too_many_arguments)]
+    async fn assert_tokens_equal(
+        session: &Session,
+        prepared: &PreparedStatement,
+        all_values_in_query_order: impl SerializeRow,
+        token_select: &PreparedStatement,
+        pk_ck_values: impl SerializeRow,
+        ks_name: &str,
+        table_name: &str,
+        pk_values: impl SerializeRow,
+    ) {
+        session
+            .execute_unpaged(prepared, &all_values_in_query_order)
+            .await
+            .unwrap();
+
+        let (value,): (i64,) = session
+            .execute_unpaged(token_select, &pk_ck_values)
+            .await
+            .unwrap()
+            .into_rows_result()
+            .unwrap()
+            .single_row::<(i64,)>()
+            .unwrap();
+        let token = Token::new(value);
+        let prepared_token = prepared
+            .calculate_token(&all_values_in_query_order)
+            .unwrap()
+            .unwrap();
+        assert_eq!(token, prepared_token);
+        let cluster_state_token = session
+            .get_cluster_state()
+            .compute_token(ks_name, table_name, &pk_values)
+            .unwrap();
+        assert_eq!(token, cluster_state_token);
+    }
+
+    // Different sizes of the key
+    {
+        session
+            .ddl("CREATE TABLE IF NOT EXISTS t1 (a text primary key)")
+            .await
+            .unwrap();
+
+        let prepared_statement = session
+            .prepare("INSERT INTO t1 (a) VALUES (?)")
+            .await
+            .unwrap();
+
+        let token_selection = session
+            .prepare("SELECT token(a) FROM t1 WHERE a = ?")
+            .await
+            .unwrap();
+
+        for i in 1..50usize {
+            eprintln!("Trying key size {}", i);
+            let mut s = String::new();
+            for _ in 0..i {
+                s.push('a');
+            }
+            let values = (&s,);
+            assert_tokens_equal(
+                &session,
+                &prepared_statement,
+                &values,
+                &token_selection,
+                &values,
+                ks.as_str(),
+                "t1",
+                &values,
+            )
+            .await;
+        }
+    }
+
+    // Single column PK and single column CK
+    {
+        session
+            .ddl("CREATE TABLE IF NOT EXISTS t2 (a int, b int, c text, primary key (a, b))")
+            .await
+            .unwrap();
+
+        // Values are given non partition key order,
+        let prepared_simple_pk = session
+            .prepare("INSERT INTO t2 (c, a, b) VALUES (?, ?, ?)")
+            .await
+            .unwrap();
+
+        let all_values_in_query_order = ("I'm prepared!!!", 17_i32, 16_i32);
+
+        let token_select = session
+            .prepare("SELECT token(a) from t2 WHERE a = ? AND b = ?")
+            .await
+            .unwrap();
+        let pk_ck_values = (17_i32, 16_i32);
+        let pk_values = (17_i32,);
+
+        assert_tokens_equal(
+            &session,
+            &prepared_simple_pk,
+            all_values_in_query_order,
+            &token_select,
+            pk_ck_values,
+            ks.as_str(),
+            "t2",
+            pk_values,
+        )
+        .await;
+    }
+
+    // Composite partition key
+    {
+        session
+        .ddl("CREATE TABLE IF NOT EXISTS complex_pk (a int, b int, c text, d int, e int, primary key ((a,b,c),d))")
+        .await
+        .unwrap();
+
+        // Values are given in non partition key order, to check that such permutation
+        // still yields consistent hashes.
+        let prepared_complex_pk = session
+            .prepare("INSERT INTO complex_pk (a, d, c, e, b) VALUES (?, ?, ?, ?, ?)")
+            .await
+            .unwrap();
+
+        let all_values_in_query_order = (17_i32, 7_i32, "I'm prepared!!!", 1234_i32, 16_i32);
+
+        let token_select = session
+            .prepare(
+                "SELECT token(a, b, c) FROM complex_pk WHERE a = ? AND b = ? AND c = ? AND d = ?",
+            )
+            .await
+            .unwrap();
+        let pk_ck_values = (17_i32, 16_i32, "I'm prepared!!!", 7_i32);
+        let pk_values = (17_i32, 16_i32, "I'm prepared!!!");
+
+        assert_tokens_equal(
+            &session,
+            &prepared_complex_pk,
+            all_values_in_query_order,
+            &token_select,
+            pk_ck_values,
+            ks.as_str(),
+            "complex_pk",
+            pk_values,
+        )
+        .await;
+    }
 }
